Add NURBS curve-surface intersection functionality
1. test/csx_test.cpp
Extend tests to verify NURBS curve-surface intersections.

Added a new set of results to test NURBS curve-surface intersection handling using a specific tolerance setting.

2. cmmcore/newthon2.h
Fix backward substitution loop index type.

Changed loop index type in backwardSubstitution function to size_t to ensure compatibility and avoid type issues.

3. cmmcore/csx.h
Implement csx function for NURBS curve-surface intersection.

Added computePtol and csx functions to handle NURBS curve-surface intersection functionality, including tolerance calculation and result storage.

+++ cmmcore/newthon2.h
@@ -193,7 +193,7 @@ constexpr void forwardSubstitution(const Matrix<N>& L, const Vector<N>& b, Vecto
 // Solve L^T * x = y
 template <size_t N>
 constexpr void backwardSubstitution(const Matrix<N>& L, const Vector<N>& y, Vector<N>& x) {
-    for (int i = static_cast<int>(N) - 1; i >= 0; --i) {
+    for (int i = N - 1; i >= 0; --i) {
         double sum = y[i];
         for (size_t k = i + 1; k < N; ++k)
             sum -= L[k][i] * x[k];
